feat(layout): added settings page and properly implemented layout
