+++ package.json
@@ -9,16 +9,16 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/node": "18.16.1",
+    "@types/node": "18.16.3",
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.1",
     "@vercel/analytics": "^1.0.0",
     "autoprefixer": "10.4.14",
     "eslint": "8.39.0",
-    "eslint-config-next": "13.3.1",
+    "eslint-config-next": "13.3.4",
     "flowbite": "^1.6.5",
     "flowbite-react": "^0.4.4",
-    "next": "13.3.1",
+    "next": "13.3.4",
     "openai": "^3.2.1",
     "postcss": "8.4.23",
     "react": "18.2.0",
@@ -28,6 +28,6 @@
   },
   "devDependencies": {
     "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.2.7"
+    "prettier-plugin-tailwindcss": "^0.2.8"
   }
 }
