+++ src/app/settings/page.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import { Alert, Badge } from 'flowbite-react';
+
+export default async function Settings() {
+	return (
+		<div className="mt-6 w-full max-w-md rounded-md bg-gray-800 p-8 shadow-lg">
+			<h1 className="mb-6 text-center text-2xl font-bold leading-tight text-white md:text-3xl">
+				🚧 Under construction 🚧
+			</h1>
+			<Alert color="warning">
+				<div>Coming soon</div>
+			</Alert>
+		</div>
+	);
+}
