+++ src/components/client/index.ts
@@ -1,0 +1,3 @@
+export * from './info-section';
+export * from './user-input';
+export * from './nav-bar';
