+++ src/components/client/info-section/assets/kofi.png
