+++ src/components/client/info-section/index.ts
@@ -1,0 +1,1 @@
+export { InfoSection } from './info-section';
