--- src/components/info-section/index.ts
@@ -1,1 +1,0 @@
-export * from './info-section';
