+++ src/components/server/loading-spinner-text/loading-spinner.tsx
@@ -1,0 +1,10 @@
+import { Spinner } from 'flowbite-react';
+
+export const LoadingSpinner = () => {
+	return (
+		<div role="status">
+			<Spinner aria-label={'button loading spinner'} />
+			Loading...
+		</div>
+	);
+};
